Move start_multiple
