+++ service/src/scripts/run_local/index.ts
@@ -2,7 +2,7 @@ import "dotenv/config";
 import { sigServer } from '@libp2p/webrtc-star-signalling-server'
 import { archLogger } from '../../utils/chalk-theme';
 import { getLocalStarSignallingPort, parseLocalArchCountArgs } from "./helpers";
-import { startMultipleLocal } from "../../start_multiple";
+import { startMultipleLocal } from "./start_multiple";
 
 const starServerPort = getLocalStarSignallingPort();
 const archCount = parseLocalArchCountArgs();
